requirements.txt
cycler==0.10.0
kiwisolver==1.1.0
matplotlib==3.1.1
numpy==1.17.3
pandas==0.25.3
Pillow==6.2.1
pkg-resources==0.0.0
pyparsing==2.4.2
python-dateutil==2.8.1
pytz==2019.3
six==1.12.0
